// archive/workspaces-app/src/types/internal.ts
export interface WorkspaceItem {
    id?: string;
    type?: "workspace";
    children: Array<RowItem | ColumnItem | GroupItem | WindowItem>;
    config?: {
        name?: string;
        context?: object;
        [k: string]: object | string;
    };
}

export interface GroupItem {
    id?: string;
    type: "group";
    children: WindowItem[];
    config?: {
        [k: string]: object;
    };
}

export interface WindowItem {
    id: string;
    type: "window";
    config: {
        url: string;
        appName: string;
        windowId?: string;
        isMaximized: boolean;
        isLoaded: boolean;
        isFocused: boolean;
        workspaceId?: string;
        frameId?: string;
        positionIndex?: number;
        title?: string;
        context?: string;
    };
}

export interface RowItem {
    id?: string;
    type: "row";
    children: Array<RowItem | ColumnItem | GroupItem | WindowItem>;
    config?: {
        [k: string]: object;
    };
}

export interface ColumnItem {
    id?: string;
    type: "column";
    children: Array<RowItem | ColumnItem | GroupItem | WindowItem>;
    config?: {
        [k: string]: object;
    };
}

export interface WorkspaceSummary {
    id: string;
    config: WorkspaceConfig;
}

export interface WorkspaceConfig {
    frameId: string;
    title: string;
    positionIndex: number;
    name: string;
    layoutName?: string;
}
// ...
export interface Window {
    id: string;
    bounds?: Bounds;
    appName?: string;
    windowId?: string;
    url?: string;
}

export interface Workspace {
    id: string;
    windows: Window[];
    layout: GoldenLayout;
    context?: object;
}

export interface WorkspaceLayout {
    name: string;
    type: "Workspace";
    metadata?: object;
    components: Array<{ type: "Workspace"; state: WorkspaceItem }>;
}